\name{wrapperEstimNCP}

\alias{wrapperEstimNCP}

\title{
wrapperEstimNCP
}

\description{
 A concise (1-5 lines) description of what the function does.
}

\usage{
wrapperEstimNCP(X, min.ncp = 0, max.ncp)
}

\arguments{
  \item{X}{Describe \code{X} here}

  \item{min.ncp}{Describe here}

  \item{max.ncp}{Describe here}
}

\details{
If necessary, more details than the description above
}

\value{
Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}

\references{
Josse, J. and Husson, F. (2012).
Selecting the number of components in PCA using cross-validation approximations. \emph{Computational Statistics and Data Analysis}, \bold{56}, 1869-1879.
}

\author{
who you are
}

\keyword{ kwd1 }
\keyword{ kwd2 }
