\name{wrapperSVD}
\alias{wrapperSVD}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
wrapperSVD
}
\description{
 A concise (1-5 lines) description of what the function does.
}
\usage{
wrapperSVD(X, row.w = NULL, col.w = NULL, ncp = Inf)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{X}{Describe \code{X} here
}
  \item{row.w}{Describe \code{row.w} here
}
  \item{col.w}{Describe \code{col.w} here
}
  \item{ncp}{Describe \code{ncp} here
}
}
\details{
 If necessary, more details than the description above
}
\value{
Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
put references to the literature/web site here
}
\author{
who you are
}
%\note{
%%  further notes
%}

%% Make other sections like Warning with \section{Warning }{....}

\seealso{
objects to See Also as %\code{\link{help}},
}
%\examples{}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ kwd1 }
\keyword{ kwd2 }% __ONLY ONE__ keyword per line
