\name{NCI60}

\alias{NCI60}

\docType{data}

\title{Data of the NCI-60 cell lines with missing rows}

\description{
The \code{NCI60} data contained transcriptomic and proteomic
expression for a collection of 60 cell lines from the National
Cancer Institute (NCI-60). Data tables with missing rows were
generated for illustration purposes.
}

\usage{data(NCI60)}

\format{
A list with three components, \code{cell.line}, \code{completeData} and
\code{incompleteData}:

  \describe{
  \item{\code{cell.line}}{character vector of cancer types:
          colon (CO), renal (RE), ovarian (OV), breast (BR),
          prostate (PR), lung (LC), central nervous system (CNS),
          leukemia (LE) and melanoma (ME).}
  }

  \code{completeData} and \code{incompleteData}, each one containing a
  list with the following components:

  \describe{
  \item{\code{trans}}{matrix containing 60 rows and 300 columns.
          The mRNA transcription levels of the NCI60 cell lines. The
          \code{trans} \code{incompleteData} table containing 12 missing
          rows.}

    \item{\code{prote}}{matrix containing 60 rows and 162 columns.
          The protein abundance levels of the NCI60 cell lines. The
          \code{prote} \code{incompleteData} table containing 8 missing
          rows.}
  }
}

\details{
The transcriptome data was retrieved directly from the
\code{NCI60_4arrays} package. This data table containe the gene
expression profiles generated by the Agilent platform with only
few hundreds of genes randomly selected to keep the size of the
Bioconductor package small. However, the full dataset is available
in Reinhold \emph{et al.} (2012).

The proteomic data was retrieved directly from the
\code{rcellminerData} package. Protein abundance levels were
available for 162 proteins.

The scripts used to generate this data are contained within
the inst/extdata folder of the \code{missRows} package.
}

\source{
\code{NCI60_4arrays} package.

\code{rcellminerData} package.
}

\references{
Reinhold W.C., Sunshine M., Liu H., Varma S., Kohn K.W., Morris J.,
Doroshow J., and Pommier Y.
Cellminer: A web-based suite of genomic and pharmacologic tools to
explore transcript and drug patterns in the NCI-60 cell line set.
\emph{Cancer Research}, \bold{72}(14):3499-511, 2012.

The CellMiner project website: http://discover.nci.nih.gov/cellminer
}

\keyword{datasets}
